+++ packages/react/src/components/index.ts
@@ -70,3 +70,5 @@ export type {
 export * from './Typography/';
 export * from './Accordion';
 export * from './NativeSelect';
+export * from './Alert';
+export * from './Tag';