+++ storefront/layouts/NavPageLayout/NavPageLayout.tsx
@@ -17,26 +17,25 @@ interface NavPageLayoutData {
 
 const NavPageLayout = ({ content, data }: NavPageLayoutProps) => {
   return (
-    <div>
-      <style>{`
-        body {
-          background: #f4f5f6;
-        }
-      `}</style>
+    <main
+      id='main'
+      className='grey-background'
+    >
       <ImageSection
-        src={data.imageSection.src}
+        imgSrc={data.imageSection.imgSrc}
+        imgAlt={data.imageSection.imgAlt}
         backgroundColor={data.imageSection.backgroundColor}
         imgHeight={220}
         imgWidth={220}
         title={data.imageSection.title}
         description={data.imageSection.description}
         link={data.imageSection.link}
-        imagePosition='right'
+        imgPosition='right'
       ></ImageSection>
       <div className={classes.content}>
         <Container>{content}</Container>
       </div>
-    </div>
+    </main>
   );
 };